+++ src/main/java/gdg/whowantit/dto/UserDto/UserRequestDto.java
@@ -1,0 +1,4 @@
+package gdg.whowantit.dto.UserDto;
+
+public class UserRequestDto {
+}
